Add files via upload
The Annealing_2000K plots show what happens to the crystal when temperature is ramped up to 2000K. It indicates a melting temperature of around 1600K, which is also what our paper reports using the BOP. However, the actual experimental melting temperature is found to be 1365K. 
The .lammps file is the new deposition file. It is not done yet but has some adjustments to the previous one, which are: The deposition Temp. is now 1000K (as reported in the paper), atoms are injected far above the surface and both Cd and Te are deposited.